move guts of reconfigurator-cli into separate package
