+++ dev-tools/repl-utils/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "omicron-repl-utils"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow.workspace = true
+camino.workspace = true
+clap.workspace = true
+reedline.workspace = true
+omicron-workspace-hack.workspace = true
